+++ identity/identity-application/build.gradle.kts
@@ -11,6 +11,7 @@ version = "0.0.1-SNAPSHOT"
 dependencies {
     implementation(project(":identity:identity-api"))
     implementation(project(":identity:identity-server:identity-server-domain"))
+    implementation(project(":identity:identity-server:identity-server-persistence"))
     implementation(project(":reservation:reservation-api"))
     implementation(project(":board:board-api"))
     implementation(project(":event-relay:event-relay-starter"))
