+++ identity/identity-server/identity-server-persistence/src/main/java/com/seatliberator/seatliberator/identity/server/persistence/account/jpa/JpaCredentialAccountStore.java
@@ -1,9 +1,9 @@
-package com.seatliberator.seatliberator.identity.infrastructure.persistence.jpa;
+package com.seatliberator.seatliberator.identity.server.persistence.account.jpa;
 
 import com.seatliberator.seatliberator.identity.application.port.out.AuthRecord;
 import com.seatliberator.seatliberator.identity.application.port.out.CredentialAccountStore;
-import com.seatliberator.seatliberator.identity.infrastructure.persistence.jpa.repository.CredentialAccountRepository;
 import com.seatliberator.seatliberator.identity.server.domain.account.CredentialAccount;
+import com.seatliberator.seatliberator.identity.server.persistence.account.jpa.repository.CredentialAccountRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
