+++ identity/identity-server/identity-server-persistence/src/main/java/com/seatliberator/seatliberator/identity/server/persistence/account/jpa/JpaUserStore.java
@@ -1,9 +1,9 @@
-package com.seatliberator.seatliberator.identity.infrastructure.persistence.jpa;
+package com.seatliberator.seatliberator.identity.server.persistence.account.jpa;
 
 
 import com.seatliberator.seatliberator.identity.application.port.out.UserStore;
-import com.seatliberator.seatliberator.identity.infrastructure.persistence.jpa.repository.UserRepository;
 import com.seatliberator.seatliberator.identity.server.domain.account.User;
+import com.seatliberator.seatliberator.identity.server.persistence.account.jpa.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
