+++ identity/identity-server/identity-server-persistence/src/main/java/com/seatliberator/seatliberator/identity/server/persistence/account/jpa/repository/FederatedAccountRepository.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.identity.infrastructure.persistence.jpa.repository;
+package com.seatliberator.seatliberator.identity.server.persistence.account.jpa.repository;
 
 import com.seatliberator.seatliberator.identity.server.domain.account.FederatedAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
